refactor(tests): move tests to tests folder
